requirements.txt
requests==2.31.0
python-dotenv==1.0.0
streamlit>=1.31,<2
openai>=1.10.0,<2
google-generativeai>=0.5.0,<1
